refactor: move fetch data func to cardDataService
