--- YouTube-client-app/src/app/youtube/models/youtube-response.model.ts
@@ -1,10 +1,0 @@
-import { Statistics } from './card-item.model';
-
-export interface YouTubeVideoStatisticsResponse {
-  items: VideoStatisticsModel[];
-}
-
-export interface VideoStatisticsModel {
-  id: string;
-  statistics: Statistics;
-}
